requirements.txt
# Core Slack bot dependencies (Python 3.13 compatible)
slack-bolt==1.21.2
slack-sdk==3.34.0

# Web framework for HTTP endpoints
fastapi==0.115.0
uvicorn[standard]==0.32.1

# Environment and configuration
python-dotenv==1.0.1

# HTTP requests
requests==2.32.3

# Database - PostgreSQL
psycopg2-binary==2.9.11
sqlalchemy==2.0.36

# Market data (Required)
finnhub-python==2.4.25

# Utilities
structlog==24.5.0
python-dateutil==2.9.0

# Note: Removed alpaca-trade-api due to Python 3.13 compatibility issues
# Will implement direct REST API calls for Alpaca integration